requirements.txt
h5py==2.10.0
tensorflow==1.12.0
keras==2.2.0
tensorflow-estimator==1.10.12
tensorflow-serving-api==1.12.0
numpy
pillow
scikit-image
opencv-python
pandas
jupyter
sagemaker